backend/server/requirements.txt
APScheduler==3.6.0
botocore==1.12.129
certifi==2018.11.29
chardet==3.0.4
colorama==0.3.9
config==0.4.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.7
django-admin==1.3.2
django-allauth==0.39.1
django-braces==1.13.0
django-cors-headers==2.4.0
django-excel-response2==2.0.8
django-heroku==0.3.1
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-six==1.0.4
djangorestframework==3.9.1
docutils==0.14
gunicorn==19.9.0
idna==2.8
jmespath==0.9.4
oauthlib==3.0.1
psycopg2==2.7.7
pyasn1==0.4.5
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python3-openid==3.1.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.2.0
requirements==0.1
rsa==3.4.2
s3transfer==0.2.0
screen==1.0.1
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
tzlocal==1.5.1
urllib3==1.24.1
whitenoise==4.1.2
xlwt==1.3.0
